+++ RELEASES.md
@@ -2,6 +2,7 @@ Version jvePilot-0.8.4(v3.3.0) (2021-XX-XX)
 ========================
 * Fix radar thanks to xps
 * Metric ACC control.  Starts at 30kph, ACC +/- changes by 5kph
+* Add support for vehicles that have been modified to allow no minimum steer speed. 
 
 Version jvePilot-0.8.4(v3.2.2) (2021-05-26)
 ========================